backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
certifi==2025.4.26
click==8.1.8
colorama==0.4.6
distro==1.9.0
dominate==2.9.1
Flask==3.1.0
Flask-Bootstrap==3.3.7.1
flask-cors==6.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
MarkupSafe==3.0.2
openai==1.82.0
pydantic==2.11.5
pydantic_core==2.33.2
pygame==2.6.1
python-dotenv==1.1.1
sniffio==1.3.1
SQLAlchemy==2.0.40
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
visitor==0.1.3
Werkzeug==3.1.3
WTForms==3.2.1
